requirements.txt
Flask==2.2.5
mysql-connector-python==8.0.33
mysqlclient==2.1.1
pandas==2.2.2
numpy==1.24.4
secrets
Werkzeug==2.2.3
plotly==5.16.1
gunicorn